Query a specific encoder to get the complete set of encoder options
During testing discovered that requesting encoder config options without specifying the actual encoder leads to a much smaller set of "universal" options. Now we provide a specific encoder -- namely the first/primary one.

This is another nail in the coffin of supporting more than one video encoder with some mechanism for selecting or changing. Any attempt in the future to do that will require pretty large overhaul of this code.