Make Onvif nodes adhere to NEPI config folder structure
Previously, we were storing the config files (without symlinks) in a new subdirectory structure, but that doesn't work properly with the config_mgr.

Ideally, this would be possible because eventually /opt/nepi/ros/etc is going to get really cluttered. But requires changes to SaveCfgIf (python and C++) and perhaps also config_mgr (though it might be as simple as passing a string with a custom path rather than the node name?)