ONVIF Mgr: Always report devices as supporting video and pan/tilt
Discovered that our previous method of looking for NetworkVideoTransmitter (video) and ptz (pan/tilt) was not universal and haven't yet figured out a universal way to determine which of these functions is supported by a given device. So for now, leaving it up to users to properly configure discovered devices per hardware capabilities.